requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.7
Django==5.0.4
django-import-export==3.3.8
et-xmlfile==1.1.0
filelock==3.12.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.3.0
platformdirs==3.11.0
python-decouple==3.8
python-gettext==5.0
PyYAML==6.0.1
sqlparse==0.5.0
tablib==3.5.0
tzdata==2023.4
virtualenv==20.24.5
virtualenvwrapper-win==1.2.7
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
django-modeltranslation== 0.19.7
